requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.11.0
PyMuPDF==1.24.2
google-generativeai
requests==2.31.0
werkzeug==3.0.1